pyproject.toml
[project]
name = "gmocoin-exec-alert"
version = "0.1.0"
description = "GMO Coin private websocket order/execution notifications -> PagerDuty alerts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "websockets>=13.0",
]

[project.scripts]
gmocoin-exec-alert = "gmocoin_exec_alert.main:main"

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

